pyproject.toml
[project]
name = "cv"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Mizerael", email = "dan-acer@outlook.com" }
]
dependencies = [
    "pillow>=10.3.0",
    "numpy>=1.24.4",
    "opencv-python>=4.9.0.80",
    "matplotlib>=3.7.5",
    "imagehash>=4.3.1",
    "tqdm>=4.66.2",
    "clip>=0.2.0",
    "torch>=2.2.2",
    "torchvision>=0.17.2",
    "transformers>=4.39.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
hello = "cv:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/cv"]
